Conflicts:
	src/main/java/smartcampus/webtemplate/controllers/NotificationController.java

+++ src/main/java/eu/trentorise/smartcampus/corsi/model/CorsoLite.java
@@ -32,7 +32,7 @@ public class CorsoLite
 		return id;
 	}
 
-	public void setId(long id)
+	public void setId(int id)
 	{
 		this.id = id;
 	}
